# Improve semanticdb-scalac performance.