--- a/scalameta/semanticdb-scalac-core/src/main/scala/scala/meta/internal/semanticdb/Inferred.scala
+++ b/scalameta/semanticdb-scalac-core/src/main/scala/scala/meta/internal/semanticdb/Inferred.scala
@@ -1,10 +1,12 @@
 package scala.meta
 package internal.semanticdb
 
-case class SyntheticRange(start: Int, end: Int, symbol: Symbol) {
+import org.langmeta.internal.semanticdb.{schema => s}
+
+case class SyntheticRange(start: Int, end: Int, symbol: String) {
   def addOffset(offset: Int) = SyntheticRange(start + offset, end + offset, symbol)
-  def toMeta(input: Input): ResolvedName =
-    ResolvedName(Position.Range(input, start, end), symbol, isDefinition = false)
+  def toSchema(input: Input): s.ResolvedName =
+    s.ResolvedName(Some(s.Position(start, end)), symbol)
 }
 case class AttributedSynthetic(text: String, names: List[SyntheticRange]) {
   def +(other: String) = AttributedSynthetic(text + other, names)
@@ -14,7 +16,7 @@ case class AttributedSynthetic(text: String, names: List[SyntheticRange]) {
 
 object AttributedSynthetic {
   val empty = AttributedSynthetic("", Nil)
-  val star = AttributedSynthetic("*", List(SyntheticRange(0, 1, Symbol("_star_."))))
+  val star = AttributedSynthetic("*", List(SyntheticRange(0, 1, Symbol("_star_.").toString())))
   def apply(text: String): AttributedSynthetic = AttributedSynthetic(text, Nil)
   def mkString(synthetics: List[AttributedSynthetic], sep: String): AttributedSynthetic =
     synthetics match {
@@ -43,7 +45,7 @@ case class Inferred(
   private def all: List[AttributedSynthetic] =
     (select :: targs :: conversion :: args :: Nil).flatten
 
-  def toSynthetic(input: Input, pos: Position): Synthetic = {
+  def toSynthetic(input: Input, pos: s.Position): s.Synthetic = {
     def onlyConversionIsDefined =
       conversion.isDefined &&
         select.isEmpty &&
@@ -60,7 +62,7 @@ case class Inferred(
       all.foldLeft(start)(_ + _)
     }
     val syntheticInput = Input.Synthetic(synthetic.text, input, pos.start, pos.end)
-    val names = synthetic.names.map(_.toMeta(syntheticInput))
-    new Synthetic(pos, synthetic.text, names)
+    val names = synthetic.names.map(_.toSchema(syntheticInput))
+    s.Synthetic(Some(pos), synthetic.text, names)
   }
 }