--- a/tests/jvm/src/test/scala-2.12/scala/meta/tests/semanticdb/MemberSuite.scala
+++ b/tests/jvm/src/test/scala-2.12/scala/meta/tests/semanticdb/MemberSuite.scala
@@ -26,6 +26,9 @@ class MemberSuite extends DatabaseSuite(SemanticdbMode.Slim, MemberMode.All) {
       |}
     """.stripMargin,
     """
+      |_empty_.a.AA#{
+      |  aa.
+      |}
       |_empty_.a.A#{
       |  G#
       |  F#
@@ -36,9 +39,6 @@ class MemberSuite extends DatabaseSuite(SemanticdbMode.Slim, MemberMode.All) {
       |  a.
       |  a.
       |  E#
-      |}
-      |_empty_.a.AA#{
-      |  aa.
       |}
     """.stripMargin
   )