… scalameta/semanticdb-scalac-core/src/main/scala/scala/meta/internal/semanticdb/DocumentOps.scala
… tests/jvm/src/test/scala/scala/meta/tests/semanticdb/TargetedSuite.scala