First step towards #1147. This PR also removes the scalameta dependency in the semanticdb-scalac compiler plugin, which is another first step towards making it potentially easier to upstream into scalac.

This PR refactors the compiler plugin to avoid
- -Yrangepos requirement, run scalac tokenizer directly instead
- scalameta parse step, emit resolved names directly from scalac trees
- 4 layers of converters from plugin to bytes on disk
Previously, the plugin created high-level data structures that got
converted to scalapb data structures that got converted to Array[Byte]
that got written to file.
Now, the plugin writes directly scalapb messages that get persisted
directly to disk.

By removing the scalameta parse step, a few resolved names are now missing:
- constructor references
- private[within]
- named argument parameters

The named argument parameters are the most critical issue, in my
opinion. I would like to figure out to get those back asap. However, I
believe we should try and merge this before the PR gets too big.

I will post  benchmark numbers as they appear. Early numbers hint the overhead has dropped from ~32% to 19%.