@DavidDudson 2018-01-18
Do you have any performance numbers for his change? I would like to know how fast the semanticdb really is.