@olafurpg 2018-01-18
I will post benchmark numbers as they appear @DavidDudson . Early numbers hint the overhead has dropped from ~40% to 17%.