@xeno-by 2018-01-18
@olafurpg What about the overhead of the LSP mode, where: 1) documents are thin, 2) denotations are only saved for definitions, not for references, 3) synthetics are not generated?