@olafurpg 2018-01-18
@xeno-by in LSP mode the overhead seems to be ~11% compared to scalac -Yrangepos. Note that -Yrangepos is not enabled in LSP mode because we no longer need it.

```
[info] QuickScalacRangepos.run              sample   41  1262.076 ± 21.051  ms/op
[info] QuickScalametaBaseline.run          semanticdb-s.jar  sample   37  1407.302 ± 40.350  ms/op
```