@olafurpg 2018-01-18
I think a good goal to aim for is ~3-5% overhead similar to Index-While-Building in clang, so we still have more work to do 😄 

```
Benchmark               Mode    Cnt Score
QuickScalacRangepos     sample  45  1202.880 ± 15.870 ms/op
QuickScalametaBaseline  sample  40  1349.203 ± 21.157 ms/op (12.2% overhead)
```