@olafurpg 2018-01-19
Here are the numbers comparing master and this PR, I ran the benchmarks both with -Yrangepos enabled and disabled for scalameta since the compiler plugin no longer requires -Yrangepos 

```// Master
Benchmark                      Mode    Cnt Score
QuickScalacBaseline            sample  48  1093.927 ± 20.610 ms/op
QuickScalacRangepos            sample  45  1133.068 ± 16.021 ms/op (3.7%)
QuickScalametaBaseline         sample  36  1448.724 ± 29.351 ms/op (32%)
QuickScalametaFullContents     sample  35  1495.988 ± 23.015 ms/op
QuickScalametaFullDenotations  sample  35  1573.523 ± 27.856 ms/op
QuickScalametaFullSynthetics   sample  35  1592.038 ± 30.538 ms/op (45%)


// This PR, with -Yrangepos enabled for scalameta
Benchmark                      Mode    Cnt Score
QuickScalacBaseline            sample  50  1043.711 ± 12.039 ms/op
QuickScalacRangepos            sample  48  1111.927 ± 15.659 ms/op (6%)
QuickScalametaBaseline         sample  40  1302.594 ± 18.033 ms/op (24%)
QuickScalametaFullContents     sample  40  1294.205 ± 17.843 ms/op
QuickScalametaFullDenotations  sample  39  1395.789 ± 17.234 ms/op
QuickScalametaFullSynthetics   sample  40  1382.443 ± 24.494 ms/op (32%)

// This PR, with -Yrangepos disabled for scalameta
Benchmark                      Mode    Cnt Score
QuickScalacBaseline            sample  49  1076.695 ± 47.137 ms/op
QuickScalacRangepos            sample  45  1155.135 ± 64.021 ms/op (7.3%)
QuickScalametaBaseline         sample  41  1281.564 ± 76.211 ms/op (19%)
QuickScalametaFullContents     sample  41  1280.848 ± 70.359 ms/op
QuickScalametaFullDenotations  sample  40  1369.912 ± 60.247 ms/op (27%)
QuickScalametaFullSynthetics   sample  40  1374.106 ± 55.381 ms/op
```

The difference was less dramatic than what I had hoped for. It would be interesting to see flamegraphs to learn more where the overhead is coming from. I suspect IO.