@olafurpg 2018-01-19
I noticed that the numbers are least stable in the last benchmark, which makes sense since I was actively using the computer while running those. The other two benchmarks were run while I was asleep :)