@xeno-by 2018-01-19
@olafurpg Here are some experiments that may be interesting:
  1) Instead of traversing ASTs in a separate phase, compute ResolvedNames directly in typedIdent. That's what we do in Rsc.
  2) Instead of creating intermediate ScalaPB data structures, directly write bytes to in-memory buffers.
  3) Instead of writing to multiple files, write to one big file.
  4) Instead of using huge symbol names, shorten the names to "1", "2", "3", etc. (This will be an irreversible transformation without a decoding table, but it'll give an idea about the performance potential of this technique).